test: move to vittest
