+++ frontend/src/vitests/lib/components/AccountBadge.spec.ts
@@ -1,7 +1,3 @@
-/**
- * @jest-environment jsdom
- */
-
 import AccountBadge from "$lib/components/accounts/AccountBadge.svelte";
 import en from "$tests/mocks/i18n.mock";
 import {
