+++ frontend/src/vitests/lib/components/BitcoinFeeDisplayTest.svelte
